examples: Fix get_time_ns overflow
On 32-bit systems tv_nec and tv_nsec are 32 bits which
caused incorrect values to be returned.

+++ examples/eglgears/linux.c
@@ -7,5 +7,5 @@ get_time_ns(void)
 {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec * 1000000000 + ts.tv_nsec;
+    return (int64_t)ts.tv_sec * 1000000000 + (int64_t)ts.tv_nsec;
 }
